+++ src/main/java/step3/Car.java
@@ -1,5 +1,9 @@
 package step3;
 
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Random;
+
 public class Car {
     private String run;
 
@@ -13,5 +17,27 @@ public class Car {
 
     public void setRun(String run){
         this.run = run;
+    }
+
+    public static List<Car> set(int car){
+        List<Car> cars = new ArrayList<>();
+
+        for (int i = 0; i < car; i++) {
+            Car c = new Car( "-");
+            cars.add(c);
+        }
+
+        return cars;
+    }
+
+    public static void run(Car car){
+        Random random = new Random();
+
+        int run = random.nextInt(10);
+        if(run >= 4){
+            String running = car.getRun();
+            running += "-";
+            car.setRun(running);
+        }
     }
 }
