+++ src/main/java/step3/ResultView.java
@@ -1,16 +1,16 @@
 package step3;
 
-import static step3.Racing.run;
-
 import java.util.List;
 
 public class ResultView {
 
-    public void resultView(List<Car> cars){
-        for (Car car : cars) {
-            run(car);
-            System.out.println(car.getRun());
+    public static void race(List<Car> cars, int number){
+        for (int i = 0; i < number; i++) {
+            for (Car car : cars) {
+                    Car.run(car);
+                System.out.println(car.getRun());
+            }
+            System.out.println();
         }
-        System.out.println();
     }
 }
